Crate gl_generator

gl_generator

gl_generator is an OpenGL bindings generator plugin. It defines a macro named generate_gl_bindings! which can be used to generate all constants and functions of a given OpenGL version.

Example

#[phase(plugin)]
extern crate gl_generator;

use self::types::*;

generate_gl_bindings! {
    api: "gl",
    profile: "core",
    version: "4.5",
    generator: "global",
    extensions: [
        "GL_EXT_texture_filter_anisotropic",
    ],
}

Arguments

Each field can be specified at most once, or not at all. If the field is not specified, then a default value will be used.

api: The API to generate. Can be either "gl", "gles1", "gles2", "wgl", "glx", "egl". Defaults to "gl".

profile: Can be either "core" or "compatibility". Defaults to "core". "core" will only include all functions supported by the requested version it self, while "compatibility" will include all the functions from previous versions as well.

version: The requested API version. This is usually in the form "major.minor". Defaults to "1.0"

generator: The type of loader to generate. Can be either "static", "global", or "struct". Defaults to "static".

extensions: Extra extensions to include in the bindings. These are specified as a list of strings. Defaults to [].

About EGL

When you generate bindings for EGL, the following platform-specific types must be declared at the same level where you call generate_gl_bindings:

khronos_utime_nanoseconds_t

khronos_uint64_t

khronos_ssize_t

EGLNativeDisplayType

EGLNativePixmapType

EGLNativeWindowType

EGLint

NativeDisplayType

NativePixmapType

NativeWindowType

Modules

generators

registry

Functions

generate_bindings
Entry point for generating bindings based on a syntax extension invocation.