Struct gl_generator::registry::Registry

pub struct Registry {
    pub groups: Vec<Group>,
    pub enums: Vec<Enum>,
    pub cmds: Vec<Cmd>,
    pub features: Vec<Feature>,
    pub extensions: Vec<Extension>,
    pub aliases: HashMap<String, Vec<String>>,
}

Fields

groups

enums

cmds

features

extensions

aliases

Methods

impl Registry

fn from_xml<R: Reader>(data: R, ns: Ns, filter: Option<Filter>) -> Registry

Generate a registry from the supplied XML string

fn get_tys(&self) -> BTreeSet<String>

Returns a set of all the types used in the supplied registry. This is useful for working out what conversions are needed for the specific registry.

fn enum_iter<'a>(&'a self) -> EnumIterator<'a>

fn cmd_iter<'a>(&'a self) -> CmdIterator<'a>